feat: Add profile update, file management, and user status update functionalities
- Implemented ProfileUpdateRequest for user profile updates with validation.
- Created SysFileQuery for file management pagination and filtering.
- Added UserStatusUpdateRequest for updating user status with validation.
- Introduced UserEditVO for user editing view representation.
- Developed SysFileMapper for file information database interactions.
- Implemented LogRetentionScheduler for scheduled log cleanup based on configuration.
- Created ForceLogoutEvent and UserCacheRefreshEvent for user session management.
- Added ISysFileService interface for file management operations.
- Implemented SysFileServiceImpl for handling file uploads, downloads, and deletions.
- Added unit tests for ServerInfo to validate system information collection.
- Introduced CookieUtil for managing refresh token cookies in HTTP responses.

+++ quantum-common/quantum-common-orm/src/main/java/com/alpha/orm/entity/BaseEntity.java
@@ -65,6 +65,6 @@ public abstract class BaseEntity implements Serializable {
      * 用于并发控制，每次更新自动 +1
      */
     @Column(version = true)
-    private Integer version;
+    private Long version;
 
 }
